Move Float, Vec3 and other value units into stdlib, too
