+++ packages/shader-composer/src/stdlib/index.ts
@@ -5,4 +5,5 @@ export * from "./masters"
 export * from "./math"
 export * from "./rotation"
 export * from "./textures"
+export * from "./values"
 export * from "./vectors"
