+++ packages/shader-composer/src/stdlib/textures.ts
@@ -1,5 +1,6 @@
 import { $ } from "../expressions"
-import { Float, Unit, Input, Vec3, Vec4 } from "../units"
+import { Input, Unit } from "../units"
+import { Float, Vec3, Vec4 } from "./values"
 
 export const Texture2D = (sampler2D: Unit<"sampler2D">, xy: Input<"vec2">) => {
 	const unit = Vec4($`texture2D(${sampler2D}, ${xy})`, {
